requirements.txt.txt
beautifulsoup4
pandas
psycopg2-binary
requests
vaderSentiment